REF: remove remaining makers
